ingestors/file/pyproject.toml
[project]
name = "file"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [
    { name = "Dominik Fluch", email = "dominik@diefluchs.at" }
]
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.100.0",
    "uvicorn>=0.20.0",
    "Jinja2>=3.0.0",
    "python-multipart>=0.0.5",
    "pypdf>=4.0.0",
    "python-docx>=1.1.0",
    "PyYAML>=6.0"
]

[build-system]
requires = ["uv_build>=0.10.0,<0.11.0"]
build-backend = "uv_build"
